# dispatch projectChange redux action when project has not changed