### Resolves

Faster project loading.

### Proposed Changes

Make VMListenerHOC aware of the scratchGui.projectChanged state and only call onProjectChanged when the project is in the unchanged state.

### Reason for Changes

Dispatching a message through redux makes it reconsider all currently listening components. Dispatching for each block, while loading, can mean dispatching and reconsidering thousands of times if not more while deserializing a project or creating a sprite from the backpack. While dispatching is pretty quick doing this for every block ends up taking a lot of time.

### Benchmark Data

Here is some data testing the effected deserialization performance while loading 173918262.

|  device | before (ms) | after (ms) | difference (ms) | change (%) |
| --- | --- | --- | --- | --- |
|  chrome | 283 | 59 | 224 | 79.12% |
|  firefox | 496 | 137 | 359 | 72.35% |
|  safari | 458 | 65 | 393 | 85.90% |
|  chromebook | 817 | 202 | 615 | 75.32% |

### Browser Coverage
Check the OS/browser combinations tested (At least 2)

Mac
 * [x] Chrome 
 * [x] Firefox 
 * [x] Safari
 
Chromebook
 * [x] Chrome